Fully implement Player vs. Player
Most of the game state is now controlled in the game class instead of
the board class. Control in the game class makes more sense as it
controls the tic-tac-toe game. The board class just controls the state
of the board itself and checks for winning/tie arrangements.

Also, the title and end result colors were changed from green to orange.