Randomly choose parents for the next generation based on fitness
Before, two parents would be selected and their genes would be part
of the entire next generation. This caused problems because the new
population was not diverse enough.

This algorithm is a weighted random selection. Parents with a higher
fitness have a higher chance of being selected, but the same parents
won't always be chosen.